shifted decomposer bindings into its own file
